transfer event list to component
